pyproject.toml
[tool.poetry]
name = "book-service"
version = "0.1.0"
description = ""
authors = ["Artyom_Vorozhbyanov <artyom.vorozhbyanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
pillow = "^9.4.0"
flake8 = "^6.0.0"
flake8-import-order = "^0.18.2"
flake8-quotes = "^3.3.2"
flake8-builtins = "^2.1.0"
pyproject-flake8 = "^6.0.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
    exclude = [".venv", "./book/books/migrations/"]
    max-line-length = 120
    extend-ignore = ["D100", "D101", "D102", "D103", "D104"]
    per-file-ignores = """
        ./book/manage.py: Q000
    """